+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Upgraded to `twilight-model` 0.13.
+- Most types exported by this crate do not longer implement `Eq`.
+
+### Removed
+- `Number` type has been removed in twilight-model, use `f64` instead.
 
 ## [0.12.0] - 2022-07-17
 ## Added